backend/requirements.txt
fastapi
uvicorn
requests
gitpython
radon
python-dotenv
pydantic
